Track Response

+++ domain/src/main/java/com/afoxplus/domain/entities/music/Artist.kt
@@ -1,3 +1,3 @@
 package com.afoxplus.domain.entities.music
 
-data class Artist(val id: Long, val name: String, val viewUrl: String)
+data class Artist(val id: Int, val name: String, val viewUrl: String)
